Move some profile fetching code into a separate module. (#5939)
This logic can be useful independently of redux actions, for CLI scripts
which want to support profiler URLs.
