+++ src/test/store/receive-profile.test.ts
@@ -27,9 +27,9 @@ import {
   retrieveProfileOrZipFromUrl,
   retrieveProfileFromFile,
   retrieveProfilesToCompare,
-  _fetchProfile,
   retrieveProfileForRawUrl,
 } from '../../actions/receive-profile';
+import { fetchProfile as _fetchProfile } from '../../utils/profile-fetch';
 import { SymbolsNotFoundError } from '../../profile-logic/errors';
 
 import { createGeckoProfile } from '../fixtures/profiles/gecko-profile';
@@ -46,7 +46,7 @@ import {
   getProfileWithThreadCPUDelta,
 } from '../fixtures/profiles/processed-profile';
 import { getHumanReadableTracks } from '../fixtures/profiles/tracks';
-import { waitUntilState } from '../fixtures/utils';
+import { waitUntilState, extractArrayBuffer } from '../fixtures/utils';
 import { dataUrlToBytes } from 'firefox-profiler/utils/base64';
 
 import { compress } from '../../utils/gz';
@@ -86,22 +86,6 @@ function simulateSymbolStoreHasNoCache() {
       .fn()
       .mockImplementation((_debugName, _breakpadId) => Promise.resolve(null)),
   }));
-}
-
-// Returns an ArrayBuffer which contains only the bytes that
-// are covered by the Uint8Array, making a copy if needed.
-function extractArrayBuffer(bufferView: Uint8Array<ArrayBuffer>): ArrayBuffer {
-  if (
-    bufferView.byteOffset === 0 &&
-    bufferView.byteLength === bufferView.buffer.byteLength
-  ) {
-    return bufferView.buffer;
-  }
-
-  // There was extra data at the start or at the end. Make a copy.
-  const copy = new Uint8Array(bufferView.byteLength);
-  copy.set(bufferView);
-  return copy.buffer;
 }
 
 describe('actions/receive-profile', function () {
@@ -1147,193 +1131,6 @@ describe('actions/receive-profile', function () {
         phase: 'FATAL_ERROR',
         error: expect.any(Error),
       });
-    });
-  });
-
-  /**
-   * _fetchProfile is a helper function for the actions, but it is tested separately
-   * since it has a decent amount of complexity around different issues with loading
-   * in different support URL formats. It's mainly testing what happens when JSON
-   * and zip file is sent, and what happens when things fail.
-   */
-  describe('_fetchProfile', function () {
-    /**
-     * This helper function encapsulates various configurations for the type of content
-     * as well and response headers.
-     */
-    async function configureFetch(obj: {
-      url: string;
-      contentType?: string;
-      content: 'generated-zip' | 'generated-json' | Uint8Array;
-    }) {
-      const { url, contentType, content } = obj;
-      const stringProfile = serializeProfile(_getSimpleProfile());
-      const profile = JSON.parse(stringProfile);
-      let arrayBuffer;
-
-      switch (content) {
-        case 'generated-zip': {
-          const zip = new JSZip();
-          zip.file('profile.json', stringProfile);
-          arrayBuffer = await zip.generateAsync({ type: 'uint8array' });
-          break;
-        }
-        case 'generated-json':
-          arrayBuffer = encode(stringProfile);
-          break;
-        default:
-          arrayBuffer = content;
-          break;
-      }
-
-      window.fetchMock.catch(403).get(url, {
-        body: arrayBuffer,
-        headers: {
-          'content-type': contentType,
-        },
-      });
-
-      const reportError = jest.fn();
-      const args = {
-        url,
-        onTemporaryError: () => {},
-        reportError,
-      };
-
-      // Return fetch's args, based on the inputs.
-      return { profile, args, reportError };
-    }
-
-    it('fetches a normal profile with the correct content-type headers', async function () {
-      const { profile, args } = await configureFetch({
-        url: 'https://example.com/profile.json',
-        contentType: 'application/json',
-        content: 'generated-json',
-      });
-
-      const profileOrZip = await _fetchProfile(args);
-      expect(profileOrZip).toEqual({ responseType: 'PROFILE', profile });
-    });
-
-    it('fetches a zipped profile with correct content-type headers', async function () {
-      const { args, reportError } = await configureFetch({
-        url: 'https://example.com/profile.zip',
-        contentType: 'application/zip',
-        content: 'generated-zip',
-      });
-
-      const profileOrZip = await _fetchProfile(args);
-      expect(profileOrZip.responseType).toBe('ZIP');
-      expect(reportError.mock.calls.length).toBe(0);
-    });
-
-    it('fetches a zipped profile with incorrect content-type headers, but .zip extension', async function () {
-      const { args, reportError } = await configureFetch({
-        url: 'https://example.com/profile.zip',
-        content: 'generated-zip',
-      });
-
-      const profileOrZip = await _fetchProfile(args);
-      expect(profileOrZip.responseType).toBe('ZIP');
-      expect(reportError.mock.calls.length).toBe(0);
-    });
-
-    it('fetches a profile with incorrect content-type headers, but .json extension', async function () {
-      const { profile, args, reportError } = await configureFetch({
-        url: 'https://example.com/profile.json',
-        content: 'generated-json',
-      });
-
-      const profileOrZip = await _fetchProfile(args);
-      expect(profileOrZip).toEqual({ responseType: 'PROFILE', profile });
-      expect(reportError.mock.calls.length).toBe(0);
-    });
-
-    it('fetches a profile with incorrect content-type headers, no known extension, and attempts to JSON parse it it', async function () {
-      const { profile, args, reportError } = await configureFetch({
-        url: 'https://example.com/profile.file',
-        content: 'generated-json',
-      });
-
-      const profileOrZip = await _fetchProfile(args);
-      expect(profileOrZip).toEqual({ responseType: 'PROFILE', profile });
-      expect(reportError.mock.calls.length).toBe(0);
-    });
-
-    it('fails if a bad zip file is passed in', async function () {
-      const { args, reportError } = await configureFetch({
-        url: 'https://example.com/profile.file',
-        contentType: 'application/zip',
-        content: new Uint8Array([0, 1, 2, 3]),
-      });
-
-      let userFacingError;
-      try {
-        await _fetchProfile(args);
-      } catch (error) {
-        userFacingError = error;
-      }
-      expect(userFacingError).toMatchSnapshot();
-      expect(reportError.mock.calls.length).toBeGreaterThan(0);
-      expect(reportError.mock.calls).toMatchSnapshot();
-    });
-
-    it('fails if a bad profile JSON is passed in', async function () {
-      const invalidJSON = 'invalid';
-      const { args, reportError } = await configureFetch({
-        url: 'https://example.com/profile.json',
-        contentType: 'application/json',
-        content: encode(invalidJSON),
-      });
-
-      let userFacingError;
-      try {
-        await _fetchProfile(args);
-      } catch (error) {
-        userFacingError = error;
-      }
-      expect(userFacingError).toMatchSnapshot();
-      expect(reportError.mock.calls.length).toBeGreaterThan(0);
-      expect(reportError.mock.calls).toMatchSnapshot();
-    });
-
-    it('fails if a bad profile JSON is passed in, with no content type', async function () {
-      const invalidJSON = 'invalid';
-      const { args, reportError } = await configureFetch({
-        url: 'https://example.com/profile.json',
-        content: encode(invalidJSON),
-      });
-
-      let userFacingError;
-      try {
-        await _fetchProfile(args);
-      } catch (error) {
-        userFacingError = error;
-      }
-      expect(userFacingError).toMatchSnapshot();
-      expect(reportError.mock.calls.length).toBeGreaterThan(0);
-      expect(reportError.mock.calls).toMatchSnapshot();
-    });
-
-    it('fallback behavior if a completely unknown file is passed in', async function () {
-      const invalidJSON = 'invalid';
-      const profile = encode(invalidJSON);
-      const { args } = await configureFetch({
-        url: 'https://example.com/profile.unknown',
-        content: profile,
-      });
-
-      let userFacingError = null;
-      try {
-        const profileOrZip = await _fetchProfile(args);
-        expect(profileOrZip).toEqual({
-          responseType: 'PROFILE',
-          profile: extractArrayBuffer(profile),
-        });
-      } catch (error) {
-        userFacingError = error;
-      }
-      expect(userFacingError).toBeNull();
     });
   });
